Refractored the BlogActivity
