+++ app/src/main/java/tech/honeysharma/techbmechat/Blog/BlogActivity.java
@@ -9,7 +9,9 @@ import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
+import android.support.design.widget.TabLayout;
 import android.support.v4.app.NavUtils;
+import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -45,19 +47,14 @@ import tech.honeysharma.techbmechat.Utility.Utility;
 
 public class BlogActivity extends AppCompatActivity {
 
-    private RecyclerView mBlogList;
-    private LinearLayoutManager mLayoutManager;
+
     private Toolbar mToolbar;
-    private DatabaseReference mDatabase,mDatabaseUser,mDatabaseLike;
     private DatabaseReference mUserRef;
-    private FirebaseRecyclerAdapter firebaseRecyclerAdapter;
-    private ProgressDialog mProgress;
     private FirebaseAuth auth;
+    private TabLayout tabLayout;
+    private ViewPager viewPager;
+    private CustomFragmentPagerAdapter mAdapter;
 
-    private List<Blog> blogList = new ArrayList<>();
-    private List<Blog> blogs = new ArrayList<>();
-
-    private boolean mProcessLike=false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState)
@@ -65,200 +62,41 @@ public class BlogActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_blog);
 
+        auth = FirebaseAuth.getInstance();
+
         mToolbar = (Toolbar) findViewById(R.id.main_page_toolbar);
         setSupportActionBar(mToolbar);
         getSupportActionBar().setTitle("TechnoJam Chat");
 
+        tabLayout = findViewById(R.id.main_tabs);
+        viewPager = findViewById(R.id.viewPager);
+
+        mAdapter = new CustomFragmentPagerAdapter(getSupportFragmentManager());
+        viewPager.setAdapter(mAdapter);
+        tabLayout.setupWithViewPager(viewPager);
+
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         getSupportActionBar().setHomeButtonEnabled(true);
-
-
-
-        auth = FirebaseAuth.getInstance();
-        mProgress=new ProgressDialog(this);
-
-        String uid=auth.getCurrentUser().getUid();
-
-
-        mDatabaseUser= FirebaseDatabase.getInstance().getReference().child("Users");
-
-        mDatabaseLike=FirebaseDatabase.getInstance().getReference().child("Like");
-
-        mDatabase= FirebaseDatabase.getInstance().getReference().child("Blog");
 
         if (auth.getCurrentUser()!=null)
         {
             mUserRef = FirebaseDatabase.getInstance().getReference().child("Users").child(auth.getCurrentUser().getUid());
         }
 
-        mDatabaseLike.keepSynced(true);
 
-        mBlogList=(RecyclerView)findViewById(R.id.blog_list);
-        mBlogList.setHasFixedSize(true);
-        mLayoutManager = new LinearLayoutManager(this);
-        mLayoutManager.setReverseLayout(true);
-        mLayoutManager.setStackFromEnd(true);
-        mBlogList.setLayoutManager(mLayoutManager);
 
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
 
-                startActivity(new Intent(BlogActivity.this,PostActivity.class));
-                //Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                //      .setAction("Action", null).show();
-            }
-        });
+
 
     }
 
     @Override
     protected void onStart() {
         super.onStart();
-        checkConnectivity(this);
-
-        firebaseRecyclerAdapter = new FirebaseRecyclerAdapter<Blog, BlogViewHolder>
-                (Blog.class, R.layout.blog_row, BlogViewHolder.class,
-                        mDatabase) {
-
-
-            @Override
-            protected void populateViewHolder(final BlogViewHolder viewHolder, final Blog model, final int position) {
-
-
-
-                final String post_key = getRef(position).getKey();
-
-
-
-                if (FirebaseAuth.getInstance().getCurrentUser().getUid().equals(model.getUid())) {
-
-                    viewHolder.deleteIcon.setVisibility(View.VISIBLE);
-
-                } else {
-                    viewHolder.deleteIcon.setVisibility(View.INVISIBLE);
-                }
-
-
-
-
-                viewHolder.setTitle(model.getTitle());
-                viewHolder.setDesc(model.getDesc());
-                viewHolder.setImage(getApplicationContext(), model.getImage());
-                viewHolder.setUsername(model.getUsername());
-                viewHolder.setDate(model.getDate());
-
-                viewHolder.setLikeBtn(post_key);
-
-                viewHolder.mview.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        Toast.makeText(BlogActivity.this, "Post Clicked", Toast.LENGTH_LONG).show();
-
-                    }
-                });
-
-                viewHolder.deleteIcon.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-
-                        showAlertDialog(viewHolder, position, model);
-
-                    }
-                });
-
-
-                viewHolder.mLikebtn.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        mProcessLike = true;
-
-                        mDatabaseLike.addValueEventListener(new ValueEventListener() {
-                            @Override
-                            public void onDataChange(DataSnapshot dataSnapshot) {
-                                if (mProcessLike) {
-                                    if (dataSnapshot.child(post_key).hasChild(auth.getCurrentUser().getUid())) {
-                                        mDatabaseLike.child(post_key).child(auth.getCurrentUser().getUid()).removeValue();
-                                        mProcessLike = false;
-                                    } else {
-                                        mDatabaseLike.child(post_key).child(auth.getCurrentUser().getUid()).setValue(mDatabaseUser.child(auth.getCurrentUser().getUid()).child("name").toString());
-                                        mProcessLike = false;
-                                    }
-                                }
-                            }
-
-                            @Override
-                            public void onCancelled(DatabaseError databaseError) {
-
-                            }
-                        });
-                    }
-
-                });
-
-            }
-
-        };
-
-        mBlogList.setAdapter(firebaseRecyclerAdapter);
-    }
-
-    private void showAlertDialog(final BlogViewHolder viewHolder, final int position, final Blog model) {
-
-        AlertDialog.Builder builder = new AlertDialog.Builder(BlogActivity.this);
-        builder.setTitle("Delete Blog");
-        builder.setMessage("Are you sure you want to delete this blog?");
-        builder.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-
-
-                Log.e("BlogActivity", "onClick: ");
-
-                Query query = FirebaseDatabase.getInstance().getReference().child("Blog");
-                query.addListenerForSingleValueEvent(new ValueEventListener() {
-                    @Override
-                    public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-
-                        for(DataSnapshot singlesnapshot : dataSnapshot.getChildren()) {
-
-                             String key = singlesnapshot.getKey();
-
-                            Log.e("BlogActivity", "onDataChange: Model UID: " + model.getUid());
-                            Log.e("BlogActivity", "onDataChange: Model title: " + model.getTitle());
-                            Log.e("BlogActivity", "onDataChange: Model desc: " + model.getDesc());
-
-                             if (model.getUid().equals(FirebaseAuth.getInstance().getCurrentUser().getUid())
-                                     && model.getTitle().equals(singlesnapshot.getValue(Blog.class).getTitle())
-                                     && model.getDesc().equals(singlesnapshot.getValue(Blog.class).getDesc())) {
-
-                                 FirebaseDatabase.getInstance().getReference().child("Blog").child(key).removeValue();
-
-                             }
-
-                        }
-
-                    }
-
-                    @Override
-                    public void onCancelled(@NonNull DatabaseError databaseError) {
-                        Log.e("BlogActivity", "onCancelled: " + databaseError.getMessage());
-                    }
-                });
-            }
-        });
-        builder.setNegativeButton("No", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-
-                dialogInterface.dismiss();
-
-            }
-        });
-        builder.show();
 
     }
+
+
 
     @Override
     protected void onResume() {
@@ -333,109 +171,13 @@ public class BlogActivity extends AppCompatActivity {
 
     }
 
-    public static class BlogViewHolder extends RecyclerView.ViewHolder{
-
-        View mview;
-        TextView post_uname,post_date,liketext, post_title, post_desc;
-
-        ImageButton mLikebtn;
-        ImageView deleteIcon;
-
-        DatabaseReference mDatabaseLike;
-        FirebaseAuth mAuth;
-
-        public BlogViewHolder(View itemView) {
-            super(itemView);
-            mview=itemView;
-
-
-            mDatabaseLike=FirebaseDatabase.getInstance().getReference().child("Like");
-            mAuth=FirebaseAuth.getInstance();
-            mDatabaseLike.keepSynced(true);
-
-            mLikebtn=(ImageButton)mview.findViewById(R.id.likeimg);
-            liketext=(TextView)mview.findViewById(R.id.liketext);
-            post_title=(TextView)mview.findViewById(R.id.post_title);
-            post_desc=(TextView)mview.findViewById(R.id.post_desc);
-
-            deleteIcon = mview.findViewById(R.id.delete_icon);
-
-            post_uname= (TextView) mview.findViewById(R.id.post_uname);
-            post_date= (TextView) mview.findViewById(R.id.post_date);
-
-            post_uname.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    System.out.println("Clicked the Username");
-                }
-            });
-
-            post_date.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    System.out.println("Clicked the date");
-                }
-            });
-
-        }
-
-        public void setLikeBtn(final String post_key){
-            mDatabaseLike.addValueEventListener(new ValueEventListener() {
-                @Override
-                public void onDataChange(DataSnapshot dataSnapshot) {
-                    if (dataSnapshot.child(post_key).hasChild(mAuth.getCurrentUser().getUid())){
-                        mLikebtn.setImageResource(R.drawable.ic_action_like_n);
-                        liketext.setText("Dislike");
-                    }else{
-                        mLikebtn.setImageResource(R.drawable.ic_action_like);
-                        liketext.setText("Like");
-                    }
-                }
-
-                @Override
-                public void onCancelled(DatabaseError databaseError) {
-
-                }
-            });
-        }
-
-        public void setTitle(String title){
-           // TextView post_title=(TextView)mview.findViewById(R.id.post_title);
-            post_title.setText(title);
-        }
-
-
-        public void setDesc(String desc){
-          //  TextView post_title=(TextView)mview.findViewById(R.id.post_desc);
-            post_desc.setText(desc);
-        }
-
-        public void setImage(Context ctx, String image){
-            ImageView post_image=(ImageView)mview.findViewById(R.id.post_image);
-            Picasso.with(ctx).load(image).into(post_image);
-
-        }
-
-        public void setUsername(String uname){
-            //TextView post_uname=(TextView)mview.findViewById(R.id.post_uname);
-            post_uname.setText(uname);
-        }
-
-        public void setDate(String date){
-
-            //TextView post_date=(TextView)mview.findViewById(R.id.post_date);
-
-            post_date.setText(date);
-        }
-    }
-
-    public void checkConnectivity(Context context){
-        if(!Utility.isOnline(this)){
+    public void checkConnectivity(final Context context){
+        if(!Utility.isOnline(context)){
             Snackbar snackbar=Snackbar.make(findViewById(R.id.drawer),"No internet connection",Snackbar.LENGTH_INDEFINITE)
                     .setAction("Retry", new View.OnClickListener() {
                         @Override
                         public void onClick(View view) {
-                            checkConnectivity(BlogActivity.this);
+                            checkConnectivity(context);
                         }
                     });
             snackbar.show();
