+++ app/src/main/java/tech/honeysharma/techbmechat/Blog/NewsFragment.java
@@ -1,0 +1,30 @@
+package tech.honeysharma.techbmechat.Blog;
+
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import tech.honeysharma.techbmechat.R;
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class NewsFragment extends Fragment {
+
+
+    public NewsFragment() {
+        // Required empty public constructor
+    }
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_news, container, false);
+    }
+
+}
