+++ app/src/main/res/layout/activity_blog.xml
@@ -1,50 +1,33 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/drawer"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    >
+    tools:context="tech.honeysharma.techbmechat.Blog.BlogActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:id="@+id/appBarLayout">
 
-    <include layout="@layout/app_bar_layout" android:id="@+id/main_page_toolbar" />
+        <include layout="@layout/app_bar_layout" android:id="@+id/main_page_toolbar" />
+
+        <android.support.design.widget.TabLayout
+            android:id="@+id/main_tabs"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
+
+        </android.support.design.widget.TabLayout>
+
     </android.support.design.widget.AppBarLayout>
 
-    <RelativeLayout
-        android:layout_below="@+id/appBarLayout"
+    <android.support.v4.view.ViewPager
+        android:id="@+id/viewPager"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context="tech.honeysharma.techbmechat.Blog.BlogActivity">
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/appBarLayout">
 
-        <android.support.v7.widget.RecyclerView
-            android:layout_height="match_parent"
-            android:layout_width="match_parent"
-            android:id="@+id/blog_list"
-            >
-
-        </android.support.v7.widget.RecyclerView>
-
-
-        <android.support.design.widget.FloatingActionButton
-            android:id="@+id/fab"
-
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_alignParentEnd="true"
-            android:layout_alignParentRight="true"
-            android:layout_gravity="bottom|end"
-            android:layout_marginBottom="26dp"
-            android:layout_marginEnd="22dp"
-            android:layout_marginRight="22dp"
-            android:src="@drawable/ic_action_add" />
-
-    </RelativeLayout>
-
+    </android.support.v4.view.ViewPager>
 </RelativeLayout>
